refactor(compiler): finish lower pipeline migration to planner + translator
Lower now reads:

    pub fn lower(mut flat: FlatPackage) -> NirPackage {
        let plan = plan::plan(&mut flat);
        translate::translate(flat, plan)
    }

The historical in-place pipeline is gone; every TIR-mutating sub-pass
lives in `lower::plan::*`, and the TIR → NIR translator absorbs the
arms that fit a single fold.

Phase 7 — wide_int Match → if-else in the translator.
`lower::translate::wide_int` builds the if-else chain in TIR and runs
it back through `convert_expr`. The literal builders
(`create_i128_literal` / `create_u128_literal`) move to
`lower::wide_int_literal` so the globals planner can keep using them.
`lower/wide_int.rs` is deleted.

Phase 4 / 5 / 6 — boxing / globals / pattern relocate to the planner.
Each wraps `&mut FlatPackage` in a transient `TirModule` so the
existing pass internals run unchanged, then transfers data back. The
globals pass is split into `extract` (per-module init functions,
before boxing/closure) and `build_initialize_modules` (top-level
aggregator, after closure). Empty `Plan` structs are avoided — these
sub-passes mutate `flat` and return `()`; `LowerPlan` continues to
carry only `closure` / `value_copy` / `strings`.

Phase 3b (partial) — `RemainingClosuresRewriter` migrates into the
translator's `Closure` arm. The planner's "Phase 4" rewrite loop is
deleted; the translator's `convert_expr` now folds remaining `Closure`
nodes into `NirExprKind::ClosureToCanonical` wrapping a
`StructLiteral` of captures using `plan.closure.functor_infos`.
`SpecializerTransformer` still constructs `TirExprKind::ClosureToCanonical`
inside specialized callee bodies and is left in the planner for now;
completing the closure split (and the consequent
`TirExprKind::ClosureToCanonical` trim from TIR) is a follow-up.

The `Translator` struct gains `closure: &ClosurePlan` and shares the
`type_table` `Rc` so the new arms can inspect resolved types during
the walk.

`docs/compiler.md` updated to describe the planner / translator stages
of lower.
